requirements.txt
fastapi==0.115.12
langchain==0.3.23
langchain_community==0.3.21
sentence-transformers
transformers
torch
pypdf
langchain_google_genai==2.1.3
pydantic==2.11.3
pymongo==4.12.0
python-dotenv==1.1.0
qdrant_client==1.13.3
uvicorn==0.34.1
streamlit
